requirements.txt
django==5.2
djangorestframework==3.16.0
markdown==3.8
django-filter==25.1
Pillow==11.2.1